move createRouter into its own file
