+++ examples/next-prisma-todomvc/src/pages/[filter].tsx
@@ -13,8 +13,9 @@ import { ReactQueryDevtools } from 'react-query/devtools';
 import superjson from 'superjson';
 import 'todomvc-app-css/index.css';
 import 'todomvc-common/base.css';
+import { createContext } from '../server/context';
 import { inferQueryOutput, trpc } from '../utils/trpc';
-import { appRouter, createContext } from './api/trpc/[trpc]';
+import { appRouter } from './api/trpc/[trpc]';
 type Task = inferQueryOutput<'todo.all'>[number];
 
 /**
