+++ examples/next-prisma-todomvc/src/server/context.ts
@@ -1,0 +1,16 @@
+import * as trpc from '@trpc/server';
+import * as trpcNext from '@trpc/server/adapters/next';
+import { prisma } from '../utils/prisma';
+
+// create context based of incoming request
+// set as optional here so it can also be re-used for `getStaticProps()`
+export const createContext = async (
+  opts?: trpcNext.CreateNextContextOptions,
+) => {
+  return {
+    req: opts?.req,
+    prisma,
+    task: prisma.task,
+  };
+};
+export type Context = trpc.inferAsyncReturnType<typeof createContext>;
