+++ examples/next-prisma-todomvc/src/server/createRouter.ts
@@ -1,0 +1,6 @@
+import * as trpc from '@trpc/server';
+import { Context } from './context';
+
+export function createRouter() {
+  return trpc.router<Context>();
+}
